:bug: fix: 공지 상세 조회 API 열람/미열람 인원 계산 로직 수정

+++ src/main/java/umc/product/domain/notice/converter/response/NoticeConverter.java
@@ -35,7 +35,7 @@ public class NoticeConverter {
                 .build();
     }
     
-    public NoticeDetailResponse toNoticeDetailResponse(Notice notice, NoticeMember noticeMember, Integer readCount, Integer unreadCount) {
+    public NoticeDetailResponse toNoticeDetailResponse(Notice notice, NoticeMember noticeMember, Long readCount, Long unreadCount) {
         return NoticeDetailResponse.builder()
                 .noticeId(notice.getId())
                 .title(notice.getTitle())
